enable 'mk o.thread' under plan9port

+++ dat.h
@@ -9,10 +9,10 @@ typedef	struct	Imgupd Imgupd;
 #define	GSHORTB(p)	((p)[0]<<8|((p)[1]))
 #define	GLONG(p) 	((p)[0]|((p)[1]<<8)|((p)[2]<<16)|((p)[3]<<24))
 #define	GLONGB(p)	(((p)[0]<<24)|((p)[1]<<16)|((p)[2]<<8)|(p)[3])
-#define	PSHORT(p,v)	(p)[0]=(v);(p)[1]=(v)>>8
-#define	PSHORTB(p,v)	(p)[0]=(v)>>8;(p)[1]=(v)
-#define	PLONG(p,v)	(p)[0]=(v);(p)[1]=(v)>>8;(p)[2]=(v)>>16;(p)[3]=(v)>>24
-#define	PLONGB(p,v)	(p)[0]=(v)>>24;(p)[1]=(v)>>16;(p)[2]=(v)>>8;(p)[3]=(v)
+#define	PSHORT(p,v)	(p)[0]=(uchar)(v);(p)[1]=(uchar)(v)>>8
+#define	PSHORTB(p,v)	(p)[0]=(uchar)(v)>>8;(p)[1]=(uchar)(v)
+#define	PLONG(p,v)	(p)[0]=(uchar)(v);(p)[1]=(uchar)(v)>>8;(p)[2]=(uchar)(v)>>16;(p)[3]=(uchar)(v)>>24
+#define	PLONGB(p,v)	(p)[0]=(uchar)(v)>>24;(p)[1]=(uchar)(v)>>16;(p)[2]=(uchar)(v)>>8;(p)[3]=(uchar)(v)
 
 #define	MIN(x,y)		(((x) < (y)) ? (x) : (y))
